toolkit/components/extensions/test/mochitest/test_ext_cookies_permissions.html

<!DOCTYPE HTML>
<html>
<head>
  <title>WebExtension test</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SpawnTask.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/ExtensionTestUtils.js"></script>
  <script type="text/javascript" src="head.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>

<script type="application/javascript">
"use strict";

function* testCookies(options) {
  // Changing the options object is a bit of a hack, but it allows us to easily
  // pass an expiration date to the background script.
  options.expiry = Date.now() / 1000 + 3600;

  function background(options) {
    function get(details) {
      return new Promise(resolve => {
        browser.cookies.get(details, resolve);
      });
    }

    function getAll(details) {
      return new Promise(resolve => {
        browser.cookies.getAll(details, resolve);
      });
    }

    function set(details) {
      return new Promise(resolve => {
        browser.cookies.set(details, resolve);
      });
    }

    function remove(details) {
      return new Promise(resolve => {
        browser.cookies.remove(details, resolve);
      });
    }

    // Ask the parent scope to change some cookies we may or may not have
    // permission for.
    let awaitChanges = new Promise(resolve => {
      browser.test.onMessage.addListener(msg => {
        browser.test.assertEq("cookies-changed", msg, "browser.test.onMessage");
        resolve();
      });
    });

    let changed = [];
    browser.cookies.onChanged.addListener(event => {
      changed.push(`${event.cookie.name}:${event.cause}`);
    });
    browser.test.sendMessage("change-cookies");


    // Try to access some cookies in various ways.
    let { url, domain, secure } = options;

    awaitChanges.then(() => {
      return get({ url, name: "foo" });
    }).then(cookie => {
      browser.test.assertEq(options.shouldPass, cookie != null, "should pass == get cookie");

      return getAll({ domain });
    }).then(cookies => {
      if (options.shouldPass) {
        browser.test.assertEq(2, cookies.length, "expected number of cookies");
      } else {
        browser.test.assertEq(0, cookies.length, "expected number of cookies");
      }

      return Promise.all([
        set({ url, domain, secure, name: "foo", "value": "baz", expirationDate: options.expiry }),
        set({ url, domain, secure, name: "bar", "value": "quux", expirationDate: options.expiry }),
        remove({ url, name: "deleted" }),
      ]);
    }).then(() => {
      if (options.shouldPass) {
        // The order of eviction events isn't guaranteed, so just check that
        // it's there somewhere.
        let evicted = changed.indexOf("evicted:evicted");
        if (evicted < 0) {
          browser.test.fail("got no eviction event");
        } else {
          browser.test.succeed("got eviction event");
          changed.splice(evicted, 1);
        }

        browser.test.assertEq("x:explicit,x:overwrite,x:explicit,foo:overwrite,bar:explicit,deleted:explicit",
                              changed.join(","), "expected changes");
      } else {
        browser.test.assertEq("", changed.join(","), "expected no changes");
      }

      browser.test.notifyPass("cookie-permissions");
    });
  }

  let extension = ExtensionTestUtils.loadExtension({
    manifest: {
      "permissions": options.permissions,
    },

    background: `(${background})(${JSON.stringify(options)})`,
  });


  let cookieSvc = SpecialPowers.Services.cookies;

  // This will be evicted after we add a fourth cookie.
  cookieSvc.add(options.domain, "/", "evicted", "bar", options.secure, false, false, options.expiry);
  // This will be modified by the background script.
  cookieSvc.add(options.domain, "/", "foo", "bar", options.secure, false, false, options.expiry);
  // This will be deleted by the background script.
  cookieSvc.add(options.domain, "/", "deleted", "bar", options.secure, false, false, options.expiry);


  yield extension.startup();

  yield extension.awaitMessage("change-cookies");
  cookieSvc.add(options.domain, "/", "x", "y", options.secure, false, false, options.expiry);
  cookieSvc.add(options.domain, "/", "x", "z", options.secure, false, false, options.expiry);
  cookieSvc.remove(options.domain, "x", "/", false);
  extension.sendMessage("cookies-changed");

  yield extension.awaitFinish("cookie-permissions");
  yield extension.unload();


  function getCookies(host) {
    let cookies = [];
    let enum_ = cookieSvc.getCookiesFromHost(host);
    while (enum_.hasMoreElements()) {
      cookies.push(enum_.getNext().QueryInterface(SpecialPowers.Ci.nsICookie2));
    }
    return cookies.sort((a, b) => String.localeCompare(a.name, b.name));
  }

  let cookies = getCookies(options.domain);
  info(`Cookies: ${cookies.map(c => `${c.name}=${c.value}`)}`);

  if (options.shouldPass) {
    is(cookies.length, 2, "expected two cookies for host");

    is(cookies[0].name, "bar", "correct cookie name");
    is(cookies[0].value, "quux", "correct cookie value");

    is(cookies[1].name, "foo", "correct cookie name");
    is(cookies[1].value, "baz", "correct cookie value");
  } else if (options.shouldWrite) {
    // Note: |shouldWrite| applies only when |shouldPass| is false.
    // This is necessary because, unfortunately, websites (and therefore web
    // extensions) are allowed to write some cookies which they're not allowed
    // to read.
    is(cookies.length, 3, "expected three cookies for host");

    is(cookies[0].name, "bar", "correct cookie name");
    is(cookies[0].value, "quux", "correct cookie value");

    is(cookies[1].name, "deleted", "correct cookie name");

    is(cookies[2].name, "foo", "correct cookie name");
    is(cookies[2].value, "baz", "correct cookie value");
  } else {
    is(cookies.length, 2, "expected two cookies for host");

    is(cookies[0].name, "deleted", "correct second cookie name");

    is(cookies[1].name, "foo", "correct cookie name");
    is(cookies[1].value, "bar", "correct cookie value");
  }

  for (let cookie of cookies) {
    cookieSvc.remove(options.domain, cookie.name, "/", false);
  }
  // Make sure we don't silently poison subsequent tests if something goes wrong.
  is(getCookies(options.domain).length, 0, "cookies cleared");
}


add_task(function* init() {
  // We need to trigger a cookie eviction in order to test our batch delete
  // observer.
  SpecialPowers.setIntPref("network.cookie.maxPerHost", 3);
  SimpleTest.registerCleanupFunction(() => {
    SpecialPowers.clearUserPref("network.cookie.maxPerHost");
  });
});

add_task(function* test_good_cookie_permissions() {
  info('Test matching, non-secure domain with non-secure cookie');
  yield testCookies({
    permissions: ["http://example.com/", "cookies"],
    url: "http://example.com/",
    domain: "example.com",
    secure: false,
    shouldPass: true,
  });

  info('Test matching, secure domain with non-secure cookie');
  yield testCookies({
    permissions: ["https://example.com/", "cookies"],
    url: "https://example.com/",
    domain: "example.com",
    secure: false,
    shouldPass: true,
  });

  info('Test matching, secure domain with secure cookie');
  yield testCookies({
    permissions: ["https://example.com/", "cookies"],
    url: "https://example.com/",
    domain: "example.com",
    secure: true,
    shouldPass: true,
  });

  info('Test matching subdomain with superdomain privileges, secure cookie (https)');
  yield testCookies({
    permissions: ["https://foo.bar.example.com/", "cookies"],
    url: "https://foo.bar.example.com/",
    domain: ".example.com",
    secure: true,
    shouldPass: true,
  });

  info('Test matching subdomain with superdomain privileges, non-secure cookie (https)');
  yield testCookies({
    permissions: ["https://foo.bar.example.com/", "cookies"],
    url: "https://foo.bar.example.com/",
    domain: ".example.com",
    secure: false,
    shouldPass: true,
  });

  info('Test matching subdomain with superdomain privileges, non-secure cookie  (http)');
  yield testCookies({
    permissions: ["http://foo.bar.example.com/", "cookies"],
    url: "http://foo.bar.example.com/",
    domain: ".example.com",
    secure: false,
    shouldPass: true,
  });
});

add_task(function* test_bad_cookie_permissions() {
  info('Test non-matching, non-secure domain with non-secure cookie');
  yield testCookies({
    permissions: ["http://example.com/", "cookies"],
    url: "http://example.net/",
    domain: "example.net",
    secure: false,
    shouldPass: false,
    shouldWrite: false,
  });

  info('Test non-matching, secure domain with non-secure cookie');
  yield testCookies({
    permissions: ["https://example.com/", "cookies"],
    url: "https://example.net/",
    domain: "example.net",
    secure: false,
    shouldPass: false,
    shouldWrite: false,
  });

  info('Test non-matching, secure domain with secure cookie');
  yield testCookies({
    permissions: ["https://example.com/", "cookies"],
    url: "https://example.net/",
    domain: "example.net",
    secure: false,
    shouldPass: false,
    shouldWrite: false,
  });

  info('Test matching subdomain with superdomain privileges, secure cookie (http)');
  yield testCookies({
    permissions: ["http://foo.bar.example.com/", "cookies"],
    url: "http://foo.bar.example.com/",
    domain: ".example.com",
    secure: true,
    shouldPass: false,
    shouldWrite: true,
  });

  info('Test matching, non-secure domain with secure cookie');
  yield testCookies({
    permissions: ["http://example.com/", "cookies"],
    url: "http://example.com/",
    domain: "example.com",
    secure: true,
    shouldPass: false,
    shouldWrite: true,
  });

  info('Test matching, non-secure host, secure URL');
  yield testCookies({
    permissions: ["http://example.com/", "cookies"],
    url: "https://example.com/",
    domain: "example.com",
    secure: true,
    shouldPass: false,
    shouldWrite: false,
  });
});

</script>

</body>
</html>
